test_app.py
#!/usr/bin/env python3
"""
Test script to verify the emotion detection app is working correctly.
"""

import os
import sys
import subprocess
import importlib.util

def test_imports():
    """Test if all required modules can be imported"""
    print("Testing imports...")
    
    required_modules = [
        'cv2',
        'deepface', 
        'numpy',
        'PIL',
        'tkinter'
    ]
    
    missing_modules = []
    
    for module in required_modules:
        try:
            if module == 'cv2':
                import cv2
                print(f"✓ OpenCV version: {cv2.__version__}")
            elif module == 'deepface':
                import deepface
                print(f"✓ DeepFace imported successfully")
            elif module == 'numpy':
                import numpy as np
                print(f"✓ NumPy version: {np.__version__}")
            elif module == 'PIL':
                from PIL import Image
                print(f"✓ Pillow imported successfully")
            elif module == 'tkinter':
                import tkinter as tk
                print(f"✓ Tkinter imported successfully")
        except ImportError as e:
            missing_modules.append(module)
            print(f"✗ {module}: {e}")
    
    if missing_modules:
        print(f"\nMissing modules: {missing_modules}")
        print("Please install missing dependencies with: pip install -r requirements.txt")
        return False
    else:
        print("✓ All required modules imported successfully!")
        return True

def test_webcam():
    """Test if webcam is accessible"""
    print("\nTesting webcam access...")
    
    try:
        import cv2
        cap = cv2.VideoCapture(0)
        
        if not cap.isOpened():
            print("✗ Webcam not accessible")
            return False
        
        ret, frame = cap.read()
        if ret:
            print(f"✓ Webcam working - frame size: {frame.shape}")
            cap.release()
            return True
        else:
            print("✗ Could not read from webcam")
            cap.release()
            return False
            
    except Exception as e:
        print(f"✗ Webcam test failed: {e}")
        return False

def test_emotion_images():
    """Test if emotion images directory exists and has required files"""
    print("\nTesting emotion images...")
    
    emotion_images_dir = 'emotion_images'
    required_emotions = ['angry', 'disgust', 'fear', 'happy', 'sad', 'surprise', 'neutral']
    
    if not os.path.exists(emotion_images_dir):
        print(f"✗ Emotion images directory not found: {emotion_images_dir}")
        return False
    
    missing_images = []
    for emotion in required_emotions:
        image_path = os.path.join(emotion_images_dir, f"{emotion}.png")
        if not os.path.exists(image_path):
            missing_images.append(f"{emotion}.png")
    
    if missing_images:
        print(f"✗ Missing emotion images: {missing_images}")
        print("Run 'python setup_emotion_images.py' to create placeholder images")
        return False
    else:
        print("✓ All emotion images found!")
        return True

def test_deepface():
    """Test DeepFace emotion detection"""
    print("\nTesting DeepFace emotion detection...")
    
    try:
        import cv2
        import numpy as np
        from deepface import DeepFace
        
        # Create a simple test image (random noise)
        test_image = np.random.randint(0, 255, (100, 100, 3), dtype=np.uint8)
        
        # This might fail if no face is detected, which is expected
        try:
            result = DeepFace.analyze(test_image, actions=['emotion'], enforce_detection=False)
            print("✓ DeepFace emotion detection working")
            return True
        except Exception as e:
            print(f"⚠ DeepFace test inconclusive (expected with random image): {e}")
            print("✓ DeepFace module loaded successfully")
            return True
            
    except Exception as e:
        print(f"✗ DeepFace test failed: {e}")
        return False

def main():
    """Run all tests"""
    print("=" * 50)
    print("EMOTION DETECTION APP - TEST SUITE")
    print("=" * 50)
    
    tests = [
        ("Module Imports", test_imports),
        ("Webcam Access", test_webcam),
        ("Emotion Images", test_emotion_images),
        ("DeepFace Detection", test_deepface)
    ]
    
    results = []
    
    for test_name, test_func in tests:
        print(f"\n{test_name}:")
        print("-" * 30)
        result = test_func()
        results.append((test_name, result))
    
    # Summary
    print("\n" + "=" * 50)
    print("TEST SUMMARY")
    print("=" * 50)
    
    passed = 0
    total = len(results)
    
    for test_name, result in results:
        status = "PASS" if result else "FAIL"
        print(f"{test_name}: {status}")
        if result:
            passed += 1
    
    print(f"\nOverall: {passed}/{total} tests passed")
    
    if passed == total:
        print("\n🎉 All tests passed! Your app is ready to run.")
        print("\nTo start the app:")
        print("  python emotion_detector.py")
        print("\nTo set up emotion images:")
        print("  python setup_emotion_images.py")
    else:
        print("\n⚠ Some tests failed. Please fix the issues above before running the app.")
    
    return passed == total

if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)
